+++ admin/numbers.php
@@ -10,7 +10,10 @@ select
   count(id) as total,
   sum(at_door) as at_door_in,
   sum(IF(at_door=0 AND checked_in=1, 1, 0)) as pre_reg_in,
-  sum(IF(at_door=0 AND checked_in=0, 1, 0)) as pre_reg_not_in
+  sum(IF(at_door=0 AND checked_in=0, 1, 0)) as pre_reg_not_in,
+  sum(IF(badge_reprints > 0, 1, 0)) as attendees_reprinted, 
+  sum(badge_reprints) as total_reprints,
+  sum(IF(override_price = 0, 1, 0)) as comped_badges
 from attendees
 SQL
 )[0];
@@ -39,6 +42,19 @@ group by admission_level
 order by admission_level
 SQL
 );
+
+$upgrades = db_query(<<<SQL
+select 
+  original_admission_level, 
+  admission_level, 
+  count(admission_level) as level_count 
+from attendees
+where original_admission_level != admission_level
+group by original_admission_level, admission_level
+order by original_admission_level, admission_level
+SQL
+);
+
 ?>
 
 <div class="container">
@@ -94,6 +110,40 @@ SQL
               <td class="text-right"><?=$row["pre_reg_not_in"] ?></td>
             </tr>
           <? } ?>
+        </table>
+      </div>
+      <div class="col-md-6">
+        <h2>Upgrades</h2>
+        <table class="table table-bordered lead">
+          <tr>
+            <th>From</th>
+            <th>To</th>
+            <th>Total</th>
+          </tr>
+          <? foreach($upgrades as $row){?>
+            <tr>
+              <td><?=reg_level($row["original_admission_level"]) ?></td>
+              <td><?=reg_level($row["admission_level"]) ?></td>
+              <td class="text-right"><?=$row["level_count"] ?></td>
+            </tr>
+          <? } ?>
+        </table>
+      </div>
+      <div class="col-md-6">
+        <h2>Other Numbers</h2>
+        <table class="table table-bordered lead">
+          <tr>
+            <td>Comped Badges</td>
+            <td><?=$totals["attendees_reprinted"] ?></td>
+          </tr>
+          <tr>
+            <td>Attendees Needing Badge Reprinted</td>
+            <td><?=$totals["total_reprints"] ?></td>
+          </tr>
+          <tr>
+            <td>Total Badge Reprints</td>
+            <td><?=$totals["comped_badges"] ?></td>
+          </tr>
         </table>
       </div>