+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Kairos 1.2 (Aug 21, 2015)
+
+Adding badge reprinting and audit logging. 
+
+* Added badge reprinting button.
+* Added upgraded badge, reprinted badge, and comped badge numbers to numbers page.
+* Added a dollar sign to the overprice price field in the edit and pay forms. 
+* Added badge name to check in form.
+* Added attendee_logs table which records every insert and update to the attendees table. 
 
 # Kairos 1.1 (Aug 11, 2015)